Crate ncg_min

Implementation of a nonlinear conjugate gradient (NCG) method

Nonlinear conjugate gradient methods are used for numerical minimization of general functions of multiple variables by using only the function values and gradient.

This crate provides a simple implementation in Rust.

Currently implemented methods:

steepest descent (for testing)

CG_DESCENT: based on Hager & Zhang '06

Simple example

let ncg = ncg_min::NonlinearCG::new();

let r = ncg.minimize(&1f64, |x, grad| { *grad = 2. * x; x * x }).unwrap();

assert!(r.abs() <= 1e-3);

Modules

secant2
Implementation of the secant2 line minimization method by Hager & Zhang'06.

Structs

NonlinearCG
Implementation of a nonlinear conjugate gradient method.

NonlinearCGIteration
Information about a performed iteration of the nonlinear CG method

Rn
An implementation of the Lin trait: an n-dimensional real vector.

Enums

NonlinearCGError

NonlinearCGMethod

Traits

Lin
Trait defining basic operations for an element of a linear space.