Struct ncg_min::secant2::Secant2

pub struct Secant2<S: Float> {
    pub delta: S,
    pub sigma: S,
    pub epsilon: S,
    pub theta: S,
    pub rho: S,
    pub max_iter: i32,
    pub ubracket_max_iter: i32,
    pub init_bracket_max_iter: i32,
}

Implementation of the secant2 line minimization method by Hager & Zhang'06.

Fields

delta

delta for Wolfe condition

sigma

sigma for Wolfe condition

epsilon

epsilon for approximate Wolfe condition (to allow for value function to increase because of rounding of errors when close to the minimum)

theta

Bisection coefficient when secant fails; allowed values in (0, 1) (0.5 is the midpoint of the interval)

rho

Extension factor for finding the initial bracketing interval; > 1

max_iter

Maximum number of iterations

ubracket_max_iter

Maximum number of U3a--b bracketing iterations

init_bracket_max_iter

Maximum number of initial bracketing iterations

Methods

impl Secant2<f32>

fn new() -> Self

impl Secant2<f64>

fn new() -> Self

impl<S: Float> Secant2<S>

fn find_wolfe<Func>(&self, c: S, f: Func, hint: Option<(S, S)>) -> Result<S, Secant2Error> where Func: FnMut(S) -> (S, S)

Find an approximate minimum of a function ϕ satisfying the Wolfe condition.

f should be the function f = |x| (ϕ(x), ϕ'(x)).

c specifies the initial search interval (0, c).

hint may contain the value (ϕ(0), ϕ'(0)) to avoid unnecessary evalution if this value is already known (as is the case of the NonlinearCG minimization method).