Struct ncg_min::NonlinearCG

pub struct NonlinearCG<S: Float> {
    pub method: NonlinearCGMethod<S>,
    pub line_method: Secant2<S>,
    pub alpha0: S,
    pub psi2: S,
    pub grad_norm_tol: S,
    pub max_iter: i32,
}

Implementation of a nonlinear conjugate gradient method.

Fields

method

Nonlinear CG method

line_method

Parameters for line minimization secant2 method

alpha0

Initial line minimization bracketing interval: [0, alpha0]

psi2

Multiplier for initial line minimization bracketing interval: [0, psi2 * alpha], where alpha was obtained in previous iteration.

grad_norm_tol

Desired norm of the gradient

max_iter

Maximum number of iterations to take

Methods

impl NonlinearCG<f32>

fn new() -> Self

Defaults for f32 type: values mostly based on [HZ'06]

impl NonlinearCG<f64>

fn new() -> Self

Defaults for f64 type: values mostly based on [HZ'06]

impl<S: Float> NonlinearCG<S>

fn minimize<Func, V>(&self, x0: &V, f: Func) -> Result<V, NonlinearCGError<V>> where Func: FnMut(&V, &mut V) -> S, V: Lin<S=S> + Clone

Mininimize the given nonlinear function over a linear space.

The function f must provide its value as well as its gradient, returned in the provided &mut V (to avoid allocation). x0 is used as the initial guess.

fn minimize_with_trace<Func, V, Callback>(&self, x0: &V, f: Func, callback: Callback) -> Result<V, NonlinearCGError<V>> where Func: FnMut(&V, &mut V) -> S, V: Lin<S=S> + Clone, Callback: FnMut(&V, NonlinearCGIteration<V::S>)

The same as minimize, but allows to pass in a callback function that is called after every iteration. It is provided with the new point after the iteration is finished, and with additional information about the performed iteration.