Trait ncg_min::Lin

pub trait Lin {
    type S: Float;
    fn dot(&self, other: &Self) -> Self::S;
    fn scale(&mut self, a: Self::S) -> &mut Self;
    fn ray_to(&mut self, other: &Self, t: Self::S) -> &mut Self;
    fn origin(&self) -> Self;

    fn norm(&self) -> Self::S { ... }
    fn norm_squared(&self) -> Self::S { ... }
    fn normalize(&mut self) -> &mut Self { ... }
    fn dist(&self, other: &Self) -> Self::S { ... }
    fn add_mut(&mut self, other: &Self) -> &mut Self { ... }
    fn combine(&mut self, a: Self::S, other: &Self, b: Self::S) -> &mut Self { ... }
    fn project_on(&mut self, dir: &Self) -> &mut Self { ... }
    fn project_ortho(&mut self, dir: &Self) -> &mut Self { ... }
}

Trait defining basic operations for an element of a linear space.

The focus is on operations in place: methods that return a Lin object modify the object in place.

Associated Types

type S: Float

Scalars for this linear space.

Required Methods

fn dot(&self, other: &Self) -> Self::S

Dot product (inner product).

fn scale(&mut self, a: Self::S) -> &mut Self

Multiplication by a constant.

fn ray_to(&mut self, other: &Self, t: Self::S) -> &mut Self

Adds a vector multiplied by a constant to this vector.

fn origin(&self) -> Self

Return the origin of the vector space to which self belongs too.

Provided Methods

fn norm(&self) -> Self::S

Norm of the vector.

fn norm_squared(&self) -> Self::S

Square of the norm.

fn normalize(&mut self) -> &mut Self

Scale the vector so that it has a norm one.

If norm is zero, causes division by zero.

fn dist(&self, other: &Self) -> Self::S

Distance between two vectors.

Default implementation uses |x - v| = sqrt(x.x - 2 x.y + y.y) to avoid copying. Therefore it is recommended to reimplement this method.

fn add_mut(&mut self, other: &Self) -> &mut Self

Adds a vector to this vector.

fn combine(&mut self, a: Self::S, other: &Self, b: Self::S) -> &mut Self

Creates a linear combination.

fn project_on(&mut self, dir: &Self) -> &mut Self

Project on a line given by a given direction.

dir does not have to be normalized, but must be nonzero.

fn project_ortho(&mut self, dir: &Self) -> &mut Self

Project on a plane orthogonal to the given direction.

dir does not have to be normalized, but must be nonzero.